catalogo mal

+++ catalogo/src/main/java/com/example/domains/entities/FilmCategoryPK.java
@@ -16,7 +16,7 @@ public class FilmCategoryPK implements Serializable {
 	private int filmId;
 
 	@Column(name="category_id", insertable=false, updatable=false, unique=true, nullable=false)
-	private byte categoryId;
+	private int categoryId;
 
 	public FilmCategoryPK() {
 	}
@@ -26,7 +26,7 @@ public class FilmCategoryPK implements Serializable {
 	public void setFilmId(int filmId) {
 		this.filmId = filmId;
 	}
-	public byte getCategoryId() {
+	public int getCategoryId() {
 		return this.categoryId;
 	}
 	public void setCategoryId(byte categoryId) {
